fix: update font families
